Struct vm_runtime_types::value::Value

pub struct Value(_);

A Move value.

Value is just a wrapper type around [ValueImpl] which allows only Move types.

Methods

impl Value

Strategies for property-based tests using Value instances.

pub fn single_value_strategy() -> impl Strategy<Value = Self>

Returns a [Strategy] that generates random primitive (non-struct) Value instances.

pub fn nested_strategy(
depth: u32,
desired_size: u32,
expected_branch_size: u32
) -> impl Strategy<Value = Self>

Returns a [Strategy] that generates arbitrary values, including Structs.

Arguments are used for recursion and define

depth of the nested Struct

approximate max number of Values in the whole tree

number of max Values at each level

pub fn struct_strategy() -> impl Strategy<Value = Self>

Returns a [Strategy] that generates random Struct instances.

impl Value

pub fn u64(value: u64) -> Self

Return a Value representing a u64 in the VM.

pub fn address(address: AccountAddress) -> Self

Return a Value representing an AccountAddress in the VM.

pub fn bool(value: bool) -> Self

Return a Value representing an bool in the VM.

pub fn byte_array(bytearray: ByteArray) -> Self

Return a Value representing a ByteArray in the VM.

pub fn string(s: VMString) -> Self

Return a Value representing a String in the VM.

pub fn struct_(s: Struct) -> Self

Return a Value representing a Struct in the VM.

pub fn reference(reference: Reference) -> Self

Return a Value representing a Reference in the VM.

pub fn global_ref(reference: GlobalRef) -> Self

Return a Value representing a GlobalRef in the VM.

pub fn native_struct(v: NativeStructValue) -> Self

pub fn value_as<T>(self) -> VMResult<T> where
VMResult<T>: From<Value>,

Convert a Value into a T if the value represents a type T.

pub fn equals(&self, v2: &Value) -> VMResult<bool>

Eq bytecode

pub fn not_equals(&self, v2: &Value) -> VMResult<bool>

Neq bytecode

pub fn is_global_ref(&self) -> bool

pub fn size(&self) -> AbstractMemorySize<GasCarrier>

pub fn as_struct_ref(&self) -> Option<&Struct>

pub fn pretty_string(&self) -> String

impl Value

pub fn simple_serialize(&self) -> Option<Vec<u8>>

Serialize this value using SimpleSerializer.

pub fn simple_deserialize(blob: &[u8], resource: StructDef) -> VMResult<Value>

Deserialize this value using lcs::Deserializer and a provided struct definition.